+++ config/project.conf.js
@@ -29,7 +29,12 @@ export default {
 		shortCut:false,
 		tabSize:2,
 		externalLink: {
-			markdown_css:()=>'/static/css/github-markdown3.0.css'
+			markdown_css:()=>'/static/css/github-markdown3.0.css',
+      hljs_js:()=>'https://cdn.bootcdn.net/ajax/libs/highlight.js/9.12.0/highlight.min.js',
+      hljs_lang:lang=>`https://cdn.bootcdn.net/ajax/libs/highlight.js/9.12.0/languages/${lang}.min.js`,
+      hljs_css:(css)=>`https://cdn.bootcdn.net/ajax/libs/highlight.js/9.12.0/styles/github.min.css`,
+      katex_js:()=>'https://cdn.bootcdn.net/ajax/libs/KaTeX/0.11.1/katex.min.js',
+      katex_css:()=>'https://cdn.bootcdn.net/ajax/libs/KaTeX/0.11.1/katex.min.css'
 		}
 	},
 
@@ -72,7 +77,12 @@ export default {
 			preview: true,
 		},
 		externalLink: {
-			markdown_css:()=>'/static/css/github-markdown3.0.css'
+			markdown_css:()=>'/static/css/github-markdown3.0.css',
+      hljs_js:()=>'https://cdn.bootcdn.net/ajax/libs/highlight.js/9.12.0/highlight.min.js',
+      hljs_lang:lang=>`https://cdn.bootcdn.net/ajax/libs/highlight.js/9.12.0/languages/${lang}.min.js`,
+      hljs_css:(css)=>`https://cdn.bootcdn.net/ajax/libs/highlight.js/9.12.0/styles/github.min.css`,
+      katex_js:()=>'https://cdn.bootcdn.net/ajax/libs/KaTeX/0.11.1/katex.min.js',
+      katex_css:()=>'https://cdn.bootcdn.net/ajax/libs/KaTeX/0.11.1/katex.min.css'
 		}
 	},